captcha: VKBrowserProfile.CapturedAt int64 → float64 to match Swift TimeInterval
Symptom from vpn.wifi.5.log 2026-05-08:

  [Go] pow: vk_profile.json parse failed:
       json: cannot unmarshal number 1778219398.3036308
       into Go struct field VKBrowserProfile.captured_at of type int64
  [Go] pow: no captured browser profile, using generated browser_fp+device

Root cause: Swift's VKProfileEntry.captured_at is TimeInterval (Double),
serializing to JSON as a fractional number. Go's int64 unmarshal won't
coerce a float — strict-mode error returns nil from loadSavedVKProfile,
fallback path uses generated profile, all the JS-side capture work
silently produces no behavioural change.

Match Swift's type. solveCaptchaPoW age-calc gains an explicit
float64(time.Now().Unix()) cast.

Validated in vpn.wifi.6.log 2026-05-08:
  pow: using captured browser profile (browser_fp=32c, device=373c, captured 0.0 days ago)
  pow: captchaNotRobot.check response: {"status":"OK","success_token":"..."}

PoW success rate climbed from 6% (build 51 baseline, generated profile)
to 67% in vpn.wifi.7.log build 59 fresh start (4 OK / 6 attempts,
4/5 slots filled, no user-visible captcha). Captured profile is
finally being used end-to-end.

+++ pkg/proxy/captcha_pow.go
@@ -41,10 +41,14 @@ var captchaPowProfile BrowserProfile
 // string. Both go directly into our outgoing componentDone/check
 // requests without re-encoding.
 type VKBrowserProfile struct {
-	Device     string `json:"device"`
-	BrowserFp  string `json:"browser_fp"`
-	UserAgent  string `json:"user_agent"`
-	CapturedAt int64  `json:"captured_at"`
+	Device    string  `json:"device"`
+	BrowserFp string  `json:"browser_fp"`
+	UserAgent string  `json:"user_agent"`
+	// CapturedAt is unix seconds with sub-second fraction. Stored as
+	// float64 because Swift writes TimeInterval (Double) and Go's
+	// json.Unmarshal won't coerce a float into int64. Match Swift's
+	// type to keep the round-trip lossless.
+	CapturedAt float64 `json:"captured_at"`
 }
 
 // vkProfilePath holds the App Group container path to vk_profile.json.
@@ -401,7 +405,7 @@ func callCaptchaNotRobotAPI(ctx context.Context, client *http.Client, sessionTok
 	deviceParam := url.QueryEscape(string(deviceBytes))
 
 	if saved := loadSavedVKProfile(); saved != nil {
-		ageDays := float64(time.Now().Unix()-saved.CapturedAt) / 86400.0
+		ageDays := (float64(time.Now().Unix()) - saved.CapturedAt) / 86400.0
 		log.Printf("pow: using captured browser profile (browser_fp=%dc, device=%dc, captured %.1f days ago)",
 			len(saved.BrowserFp), len(saved.Device), ageDays)
 		browserFp = saved.BrowserFp
